# use ansi-html instead of ansi-to-html, ~11x faster